database: Move the connection into a separate file.
